1.加入故障检测与故障上报功能 2.完善通信协议和修复初始轮子不抱死问题

+++ HARDWARE/ENCODER/Encoder.c
@@ -145,10 +145,10 @@ void Get_Encoder(void)
 	Get_Encoder_T5();  //three
 	
 	//计算当前速度
-  LeftWheel.NowSpeed =(int)(GetEncoder.V3/SPEED_TO_ENCODER);
-	RightWheel.NowSpeed=(int)(GetEncoder.V1/SPEED_TO_ENCODER);
-	ThreeWheel.NowSpeed=(int)(GetEncoder.V5/SPEED_TO_ENCODER);
-	FourWheel.NowSpeed =(int)(GetEncoder.V2/SPEED_TO_ENCODER);
+  LeftWheel.NowSpeed =(s16)(GetEncoder.V3/SPEED_TO_ENCODER);
+	RightWheel.NowSpeed=(s16)(GetEncoder.V1/SPEED_TO_ENCODER);
+	ThreeWheel.NowSpeed=(s16)(GetEncoder.V5/SPEED_TO_ENCODER);
+	FourWheel.NowSpeed =(s16)(GetEncoder.V2/SPEED_TO_ENCODER);
 }
